Prefer abstract base classes to NotImplementedError
Provides a formal definition for classes that need to be extended,
Abstract Base Classes.

For support with python <3.4, the metaclass ABCMeta must be used. Python
3.4>= the helper class ABC is available.

The API of the affected classes does not change. But the behaviour of
child classes does. A child class that does not implement all of the
abstract class methods will fail diferently prior and after this commit.

This change may affect implementations that relied on the previous
behaviour. Either the implementations should explicitly raise
NotImplementedError (which is better than the base class doing it
implicitly), or the method is declared in the wrong class. In other
words someone did something wrong.

In any case, attempting to use a class that does not implement a method
required from its base class will fail in the following maner:

Prior to this commit:: NotImplementedError is raised when attempting to
call the method. Returning HTTP 500 to the client.

After this commit:: TypeError is raised (stating the missing attributes)
when attempting to instantiate the class. The server is unable to start.

Returning HTTP 500 is never a good thing. For that at least, I believe
this a better approach.